requirements.txt
transformers
datasets
torchaudio
librosa